ml_project/requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.1.0
backcall==0.2.0
contourpy==1.0.6
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Faker==15.2.0
fonttools==4.38.0
hydra-core==1.2.0
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
missingno==0.5.1
nest-asyncio==1.5.6
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyGObject==3.42.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
stack-data==0.6.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
